# web/config/openapi.json
{"openapi":"3.1.0","info":{"title":"utoipa-actix-web","description":"Utoipa's actix-web bindings for seamless integration of the two","contact":{"name":"Juha Kukkonen","email":"juha7kukkonen@gmail.com"},"license":{"name":"MIT OR Apache-2.0","identifier":"MIT OR Apache-2.0"},"version":"0.1.2"},"paths":{"/api/currentUser":{"get":{"summary":"Get current user","operationId":"get_current_user","responses":{"200":{"description":"Current user info","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRespone_CurrentUser"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRespone_NoLogintUser"}}}}}}},"/api/dfr/list":{"get":{"summary":"List files","operationId":"list_files","parameters":[{"name":"page_no","in":"query","description":"Page number, start from 1","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"page_count","in":"query","description":"Page count, must be greater than 0","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"min_file_size","in":"query","description":"Minimum file size","required":false,"schema":{"type":["integer","null"],"format":"int64"}},{"name":"max_file_size","in":"query","description":"Max file size","required":false,"schema":{"type":["integer","null"],"format":"int64"}},{"name":"dir_path","in":"query","description":"Dir path of the directory containing the file","required":false,"schema":{"type":["string","null"]}},{"name":"file_name","in":"query","description":"File name filtering","required":false,"schema":{"type":["string","null"]}},{"name":"file_extension","in":"query","description":"New field for file extension filtering","required":false,"schema":{"type":["string","null"]}},{"name":"file_extension_list","in":"query","description":"Optional file extension list filtering, comma(,) separated values.","required":false,"schema":{"type":["string","null"]}},{"name":"md5","in":"query","description":"MD5 hash of the file content, used for filtering files by their content.","required":false,"schema":{"type":["string","null"]}},{"name":"start_created_time","in":"query","description":"Optional time range filter for file creation.","required":false,"schema":{"type":["string","null"],"format":"date-time"}},{"name":"end_created_time","in":"query","required":false,"schema":{"type":["string","null"],"format":"date-time"}},{"name":"start_modified_time","in":"query","description":"Optional time range filter for file modification.","required":false,"schema":{"type":["string","null"],"format":"date-time"}},{"name":"end_modified_time","in":"query","required":false,"schema":{"type":["string","null"],"format":"date-time"}},{"name":"min_md5_count","in":"query","description":"Minimum file md5 count","required":false,"schema":{"type":["integer","null"],"format":"int64"}},{"name":"max_md5_count","in":"query","description":"Max file md5 count","required":false,"schema":{"type":["integer","null"],"format":"int64"}}],"responses":{"200":{"description":"The list of file info with md5 count","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileInfoList"}}}}}}},"/api/dfr/scan/start":{"post":{"summary":"Start a new file scan","operationId":"start_scan","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScanRequest"}}},"required":true},"responses":{"200":{"description":"Scan started successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RestResponse_i64"}}}},"409":{"description":"Scan already in progress"}}}},"/api/dfr/scan/stop":{"post":{"summary":"Stop the current file scan","operationId":"stop_scan","responses":{}}},"/api/login/account":{"post":{"summary":"Login user account","operationId":"login_account","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginParams"}}},"required":true},"responses":{"200":{"description":"Login result","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResult"}}}},"403":{"description":"Illegal username or password"}}}},"/api/login/captcha":{"post":{"summary":"Get captcha for login","operationId":"get_captcha","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FakeCaptchaParams"}}},"required":true},"responses":{"200":{"description":"Get captcha successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FakeCaptcha"}}}},"400":{"description":"Bad request"},"501":{"description":"Not implemented"}}}},"/api/login/outLogin":{"post":{"summary":"Logout user account","operationId":"logout_account","responses":{"200":{"description":"Logout successful"}}}},"/api/notices":{"get":{"summary":"Get notices","operationId":"get_notices","responses":{"200":{"description":"Notices","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoticeIconList"}}}},"401":{"description":"Unauthorized"}}}}},"components":{"schemas":{"CurrentUser":{"type":"object","properties":{"access":{"type":["string","null"]},"address":{"type":["string","null"]},"avatar":{"type":["string","null"]},"country":{"type":["string","null"]},"email":{"type":["string","null"]},"geographic":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Geographic"}]},"group":{"type":["string","null"]},"name":{"type":["string","null"]},"notifyCount":{"type":["integer","null"],"format":"int32","minimum":0},"phone":{"type":["string","null"]},"signature":{"type":["string","null"]},"tags":{"type":["array","null"],"items":{"$ref":"#/components/schemas/LabelKey"}},"title":{"type":["string","null"]},"unreadCount":{"type":["integer","null"],"format":"int32","minimum":0},"userid":{"type":["string","null"]}}},"FakeCaptcha":{"type":"object","properties":{"code":{"type":["integer","null"],"format":"int32","minimum":0},"status":{"type":["string","null"]}}},"FakeCaptchaParams":{"type":"object","properties":{"phone":{"type":["string","null"]}}},"FileInfo":{"type":"object","required":["inode_info","dir_path","file_path","file_name","version","scan_time"],"properties":{"dir_path":{"type":"string","description":"Dir path of the directory containing the file"},"file_extension":{"type":["string","null"],"description":"File extension"},"file_name":{"type":"string","description":"File name"},"file_path":{"type":"string","description":"File path"},"inode_info":{"$ref":"#/components/schemas/InodeInfo","description":"Inode info"},"scan_time":{"type":"string","format":"date-time","description":"scan_time is the time when the file was last scanned"},"version":{"type":"integer","format":"int64","description":"version is the version of the file, used to track changes","minimum":0}}},"FileInfoList":{"type":"object","description":"File info list with total count","required":["file_info_list","total_count"],"properties":{"file_info_list":{"type":"array","items":{"$ref":"#/components/schemas/FileInfoWithMd5Count"},"description":"file info list"},"total_count":{"type":"integer","format":"int64","description":"total file count","minimum":0}}},"FileInfoWithMd5Count":{"type":"object","description":"File info with md5 count","required":["file_info","md5_count"],"properties":{"file_info":{"$ref":"#/components/schemas/FileInfo"},"md5_count":{"type":"integer","minimum":0}}},"Geographic":{"type":"object","properties":{"city":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/LabelKey"}]},"province":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/LabelKey"}]}}},"InodeInfo":{"type":"object","description":"Inode info","required":["inode","dev_id","permissions","nlink","uid","gid","created","modified","size"],"properties":{"created":{"type":"string","format":"date-time"},"dev_id":{"type":"integer","format":"int64","description":"Device ID","minimum":0},"gid":{"type":"integer","format":"int32","minimum":0},"inode":{"type":"integer","format":"int64","description":"Inode number","minimum":0},"md5":{"type":["string","null"],"description":"File md5"},"modified":{"type":"string","format":"date-time"},"nlink":{"type":"integer","format":"int64","minimum":0},"permissions":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int64","description":"File size","minimum":0},"uid":{"type":"integer","format":"int32","minimum":0}}},"LabelKey":{"type":"object","properties":{"key":{"type":["string","null"]},"label":{"type":["string","null"]}}},"LoginParams":{"type":"object","description":"Login params","required":["username","password","autoLogin","type"],"properties":{"autoLogin":{"type":"boolean"},"password":{"type":"string"},"type":{"type":"string"},"username":{"type":"string"}}},"LoginResult":{"type":"object","required":["status","type","currentAuthority"],"properties":{"currentAuthority":{"type":"string"},"status":{"type":"string"},"type":{"type":"string"}}},"NoLogintUser":{"type":"object","required":["isLogin"],"properties":{"isLogin":{"type":"boolean"}}},"NoticeIconItem":{"type":"object","properties":{"avatar":{"type":["string","null"]},"datetime":{"type":["string","null"]},"description":{"type":["string","null"]},"extra":{"type":["string","null"]},"id":{"type":["string","null"]},"key":{"type":["string","null"]},"read":{"type":["boolean","null"]},"status":{"type":["string","null"]},"title":{"type":["string","null"]},"type":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/NoticeIconItemType"}]}}},"NoticeIconItemType":{"type":"string","enum":["notification","message","event"]},"NoticeIconList":{"type":"object","required":["total","success"],"properties":{"data":{"type":["array","null"],"items":{"$ref":"#/components/schemas/NoticeIconItem"}},"success":{"type":"boolean"},"total":{"type":"integer","format":"int32","minimum":0}}},"RestResponse_i64":{"type":"object","required":["success","code"],"properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"integer","format":"int64"},"message":{"type":["string","null"]},"success":{"type":"boolean"}}},"ScanRequest":{"type":"object","required":["scan_path"],"properties":{"scan_path":{"type":"string","description":"Scan path"}}},"UserRespone_CurrentUser":{"type":"object","required":["data","errorCode","errorMessage","success"],"properties":{"data":{"type":"object","properties":{"access":{"type":["string","null"]},"address":{"type":["string","null"]},"avatar":{"type":["string","null"]},"country":{"type":["string","null"]},"email":{"type":["string","null"]},"geographic":{"oneOf":[{"type":"null"},{"$ref":"#/components/schemas/Geographic"}]},"group":{"type":["string","null"]},"name":{"type":["string","null"]},"notifyCount":{"type":["integer","null"],"format":"int32","minimum":0},"phone":{"type":["string","null"]},"signature":{"type":["string","null"]},"tags":{"type":["array","null"],"items":{"$ref":"#/components/schemas/LabelKey"}},"title":{"type":["string","null"]},"unreadCount":{"type":["integer","null"],"format":"int32","minimum":0},"userid":{"type":["string","null"]}}},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"},"success":{"type":"boolean"}}},"UserRespone_NoLogintUser":{"type":"object","required":["data","errorCode","errorMessage","success"],"properties":{"data":{"type":"object","required":["isLogin"],"properties":{"isLogin":{"type":"boolean"}}},"errorCode":{"type":"integer","format":"int32"},"errorMessage":{"type":"string"},"success":{"type":"boolean"}}}}}}